requirements/base.txt
Django==2.0.2
django-tinymce==2.7.0
pytz==2017.3
whitenoise==3.3.1
gunicorn==19.7.1